src/mcp_client.py
"""MCP Server client using Langchain adapters."""

from typing import List, Dict, Any, Optional
import subprocess
import sys

from langchain_mcp_adapters import MCPClient, MultiServerMCPClient
from langchain_core.tools import BaseTool

from src.config import settings
from src.utils import get_logger, setup_logging

setup_logging()
logger = get_logger(__name__)


class MCPClientWrapper:
    """Wrapper for MCP Server client using Langchain adapters."""

    def __init__(self):
        """Initialize MCP client."""
        self.client: Optional[MCPClient] = None
        self.tools: List[BaseTool] = []
        self._initialize_client()

    def _initialize_client(self) -> None:
        """Initialize MCP client connection."""
        try:
            # Create MCP client with stdio transport
            # The server will be started as a subprocess
            server_command = [
                sys.executable,
                "-m",
                "src.mcp_server.server",
            ]

            self.client = MCPClient(
                transport="stdio",
                command=server_command[0],
                args=server_command[1:],
            )

            # Load tools from MCP server
            self._load_tools()

            logger.info("MCP client initialized successfully")
        except Exception as e:
            logger.error(f"Error initializing MCP client: {e}")
            raise

    def _load_tools(self) -> None:
        """Load tools from MCP server."""
        try:
            if self.client:
                # Get available tools from MCP server
                # This will be implemented based on langchain-mcp-adapters API
                # For now, we'll create a placeholder
                logger.info("Loading tools from MCP server...")
                # TODO: Implement tool loading based on actual langchain-mcp-adapters API
                self.tools = []
        except Exception as e:
            logger.error(f"Error loading MCP tools: {e}")
            raise

    def get_tools(self) -> List[BaseTool]:
        """Get available MCP tools as Langchain tools."""
        return self.tools

    async def call_tool(self, tool_name: str, arguments: Dict[str, Any]) -> Any:
        """
        Call an MCP tool.

        Args:
            tool_name: Name of the tool to call
            arguments: Tool arguments

        Returns:
            Tool result
        """
        try:
            if not self.client:
                raise RuntimeError("MCP client not initialized")

            # Call tool through MCP client
            # This will be implemented based on actual API
            logger.info(f"Calling MCP tool: {tool_name} with {arguments}")
            # TODO: Implement actual tool calling
            return None
        except Exception as e:
            logger.error(f"Error calling MCP tool {tool_name}: {e}")
            raise

    def search_documents(self, query: str, n_results: int = 5) -> List[Dict[str, Any]]:
        """
        Search documents using MCP tool.

        Args:
            query: Search query
            n_results: Number of results

        Returns:
            List of search results
        """
        # This will use the MCP tool when properly integrated
        # For now, return empty list
        logger.info(f"Searching documents with query: {query}")
        return []
