pyproject.toml
[project]
name = "aiagent-chat-with-doc"
version = "0.1.0"
description = "RAG-based conversational AI assistant with LangGraph ReAct agent, MCP Server, and Chroma vector store"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "openai>=1.12.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=0.9.0",
    "chromadb>=0.4.22",
    "streamlit>=1.32.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
